Using webpack-why plugin to modify webpack config
