+++ packages/webpack-why/test/fixtures/big.js
@@ -1,0 +1,3 @@
+require('redux')
+require('redux/lib/redux')
+require('./file')
