+++ packages/webpack-why/test/fixtures/file.js
@@ -1,0 +1,5 @@
+console.log('11111111111')
+console.log('11111111111')
+console.log('11111111111')
+console.log('11111111111')
+console.log('11111111111')
