+++ packages/webpack-why/test/fixtures/small.js
@@ -1,0 +1,1 @@
+console.log('111111111111111111111')
